[Fix] - some minor fixes of my functions.

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: aroi <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/23 14:05:52 by aroi              #+#    #+#             */
-/*   Updated: 2018/03/27 16:52:21 by aroi             ###   ########.fr       */
+/*   Updated: 2018/03/29 19:41:44 by aroi             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 
 	new_dst = (unsigned char *)dst;
 	new_src = (unsigned char *)src;
-	while ((int)n-- > 0)
+	while (n-- > 0)
 	{
 		*new_dst++ = *new_src++;
 		if (*(new_src - 1) == (unsigned char)c)
